Renamed class Property to PropertyData och changed string Value to object Value.  Added HandleXML read/write. - Andreas Brunnstrom

+++ src/FileConverter/Models/UnitPrice.cs
@@ -1,13 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace FileConverter.Models
 {
     [Serializable()]
     public class UnitPrice
     {
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Currency { get; set; }
     }
 }
